+++ TouchGFX/gui/include/gui/screen2_screen/Screen2Presenter.hpp
@@ -31,7 +31,7 @@ public:
     virtual ~Screen2Presenter() {};
 
     void setTestVal1(float newTestVal1);
-    void setTestVal2(uint16_t newTestVal2);
+    void setTestVal2(float newTestVal2);
     void setDutyCycle(int newDutyCycle);
 
     void askForDutyUp();
