+++ TouchGFX/gui/src/screen2_screen/Screen2Presenter.cpp
@@ -26,7 +26,7 @@ void Screen2Presenter::setTestVal1(float newTestVal1)
 	view.updateTestVal1(newTestVal1);
 }
 
-void Screen2Presenter::setTestVal2(uint16_t newTestVal2)
+void Screen2Presenter::setTestVal2(float newTestVal2)
 {
 	view.updateTestVal2(newTestVal2);
 }
